move reporting stuff to the reporting blueprint
