+++ templates/index.html
@@ -7,7 +7,7 @@
         <ul>
 	    {% if current_user.has_role("admin") %}
             <li>
-                <a href="/dashboard">Dashboard</a>
+                <a href="/reporting/dashboard">Dashboard</a>
             </li>
             <li>
                 <a href="/push_notifications">Push Notifications</a>
