+++ templates/reporting/dashboard_user_balance.html
@@ -3,8 +3,8 @@
 {% block content %}
 
 <a class="btn btn-primary" href="/"><i class="glyphicon glyphicon-chevron-left"></i> Back</a>
-<a class="btn btn-primary" href="/dashboard"><i class="glyphicon glyphicon-chevron-left"></i>&nbspDashboard</a>
-<a class="btn btn-primary" href="/dashboard_user_balance"><i class="glyphicon glyphicon-chevron-left"></i>&nbspUser Balance</a>
+<a class="btn btn-primary" href="dashboard"><i class="glyphicon glyphicon-chevron-left"></i>&nbspDashboard</a>
+<a class="btn btn-primary" href="dashboard_user_balance"><i class="glyphicon glyphicon-chevron-left"></i>&nbspUser Balance</a>
 
 <div class='some-page-wrapper'>
   <div class='row'>
