+++ templates/reporting/dashboard_waves.html
